refactor(server): extract HttpEndpoint host builder to its own module (ARN-208)
Address PR review (Greptile P2s): move build_http_endpoint_wasm_host from
router.rs into a dedicated http_endpoint_host.rs with module-level use imports
(replacing the fully-qualified paths). router.rs drops back under 1000 lines, so
the readability-baseline PROD_FILES_GT1000 bump is reverted (22). Pure
relocation — no behavioral change; the dispatch path calls the same helper.
