+++ .ci/readability-baseline.env
@@ -2,7 +2,7 @@
 PROD_RS_TOTAL=523
 PROD_FILES_GT300=209
 PROD_FILES_GT500=84
-PROD_FILES_GT1000=23
+PROD_FILES_GT1000=22
 PROD_MAX_FILE_LINES=2829
 PROD_MAX_FILE_PATH=crates/temper-server/src/storage/mod.rs
 ALLOW_CLIPPY_COUNT=35
